refactor(recovery): extract codex-rollout persist-outcome handling to restore_persist_outcome.rs (giant baseline 2335)
