+++ ARCHITECTURE.md
@@ -523,6 +523,7 @@ src/
 │   │   │   ├── phase_policy.rs
 │   │   │   ├── rebind_runtime.rs
 │   │   │   ├── restore_inflight.rs
+│   │   │   ├── restore_persist_outcome.rs
 │   │   │   ├── routing_orphan.rs
 │   │   │   ├── runtime.rs
 │   │   │   ├── state_extractors.rs
